Add files via upload
Project was completed when I uploaded it.
But in this there are many things I would like to change and revisit.
the use of textarea is almost dumb in my opinion, because I can use contenteditable on a div and set it to true, instead.
The div will be better when it comes to listening to the CSS I am telling it to follow. textarea just does it's own thing.
But my aim right now is to complete the "course" and upload the completed projects.

cool thing  in this was to see how localstorage was used and abused to get what we needed.